refactor: extract hook logic from BoardsDetail component
