--- trip-talk/src/app/boards/new/styles.module.css
@@ -1,81 +1,0 @@
-.form {
-  @apply flex flex-col items-start gap-10 w-full;
-}
-
-.form_title {
-  @apply text-xl font-bold text-black;
-}
-
-.form > div {
-  @apply flex w-full pb-10 border-b border-solid border-x-0 border-t-0 border-lightGray;
-}
-
-.author_info_wrapper {
-  @apply gap-10 border-none;
-}
-
-.author_info_wrapper > * {
-  flex: 1;
-}
-
-.title_wrapper {
-  @apply flex-col;
-}
-
-.content_wrapper {
-  @apply flex-col;
-}
-
-.content_wrapper > div {
-  @apply flex items-start gap-1;
-}
-
-.content_wrapper > div > p {
-  @apply text-base font-medium text-darkGray;
-}
-
-.content_wrapper > div > b {
-  @apply text-base font-medium text-lightCoral;
-}
-
-.content_wrapper > textarea {
-  @apply outline-none w-full flex py-3 px-4 rounded-lg border border-solid bg-white resize-none;
-  height: 21rem;
-  border-color: #d4d3d3;
-}
-
-.content_wrapper > textarea::placeholder {
-  color: #ababab;
-}
-
-.address_wrapper {
-  @apply flex-col gap-2 w-full;
-}
-
-.address_wrapper > div {
-  @apply flex flex-row gap-2 items-end;
-}
-
-.link_wrapper {
-  @apply flex-col gap-2;
-}
-
-.photo_wrapper {
-  @apply flex-col gap-2 w-full items-start;
-}
-
-.photo_wrapper > p {
-  @apply text-base font-medium text-darkGray;
-}
-
-.photo_wrapper > div {
-  @apply flex flex-row gap-4 items-start;
-}
-
-.button_wrapper {
-  @apply flex flex-row gap-4 items-center justify-end border-none;
-}
-
-.required_field {
-  @apply flex pt-2 text-base font-medium text-lightCoral;
-}
